--- src/controllers/todo/index.ts
@@ -1,8 +1,0 @@
-
-export const itsWorks = (request, response) => {
-  return response.json({ message: "It's Works!!" })
-}
-
-export const getTODO = (request, response) => {
-  return response.json({ todos: [] })
-}
