--- src/models/entity/ToDo.ts
@@ -1,11 +1,0 @@
-import { Column, Entity, PrimaryGeneratedColumn } from "typeorm";
-
-@Entity()
-export class ToDo {
-  @PrimaryGeneratedColumn()
-  id: number;
-
-  @Column({ name: 'name', type: 'varchar' })
-  name: String;
-
-}
