--- src/routes.ts
@@ -1,6 +1,0 @@
-import { getTODO, itsWorks } from "@controllers/todo"
-
-export const defineRoutes = (app) => {
-  app.get("/", itsWorks)
-  app.get("/todo", getTODO)
-}
